// apps/web/src/grpc/generated/learningmotors/pb/composer/PartialDictation.ts
// Original file: node_modules/protobufs/src/learningmotors/pb/composer/compositions.proto

export interface PartialDictation {
  partial_dictation_composition?: boolean;
  read_only?: boolean;
}

export interface PartialDictation__Output {
  partial_dictation_composition: boolean;
  read_only: boolean;
}